requirements.txt
django 
djangorestframework 
psycopg2_binary

asgiref==3.7.2
Django==4.2.5
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
asgiref==3.7.2
Django==4.2.5
djangorestframework==3.14.0
psycopg2-binary==2.9.8
pytz==2023.3.post1
sqlparse==0.4.4
